(PC-35295) refactor(BusinessModule): remove old component and FF
