+++ src/libs/firebase/firestore/types.ts
@@ -51,7 +51,6 @@ export enum RemoteStoreFeatureFlags {
   ENABLE_REPLICA_ALGOLIA_INDEX = 'enableReplicaAlgoliaIndex',
   SHOW_REMOTE_GENERIC_BANNER = 'showRemoteBanner',
   TARGET_XP_CINE_FROM_OFFER = 'targetXpCineFromOffer',
-  WIP_APP_V2_BUSINESS_BLOCK = 'wipAppV2BusinessBlock',
   WIP_ARTIST_PAGE = 'wipArtistPage',
   WIP_ARTIST_PAGE_IN_SEARCH = 'wipArtistPageInSearch',
   WIP_DISABLE_STORE_REVIEW = 'wipDisabledStoreReview',
